Add a veth recreation mechanism
When deploying with clab, the "interfaces" connecting the nodes to the
infrastructure are veth pairs.

The behavior of the linux kernel is that when a namespace is deleted,
any virtual interface (including the veth!) belonging to that network
namespace is deleted as well.

This means that if the router pod is deleted, the veth moved in the pod
(connecting a kind note to the "fabric") is deleted.

Here we monitor the deletion of the host side of the veth, and in case
it's deleted, we restore it effectively simulating what would happen in
case of a physical interface.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>