Veth: check for the pe leg by name in the namespace
When the veth is already created, LinkByIndex returns even if the link
belongs to another namespace (the perouter one in this case).

In order to understand if the pe side is already created and moved, we
check first in the namespace (as opposed to not finding in the current
ns), and if it's not there we look for the link by ID and we move it to
the right network namespace.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>