API: add a l2vni type
We add a l2vni type to offer a way to extend a l2 domain via a vxlan
overlay.

As opposed to l3vni veths which land directly in the vrf routing space
corresponding to the ip domain of a l3 VNI, here we enslave the veth leg
inside the PE router to the linux bridge which extends the l2 domain,
effectively adding a way to plug the domain.

Additionally, we expose a knob to either plug the host part of the veth
to an existing bridge, or to automatically create one, which is
particularly convenient in case we want to use the macvlan CNI to add a
pod to the l2 domain.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>