API: move from single localcidr to dual stack local cidr
We want to allow the user to define an ipv4 local cidr and an ipv6 local
cidr.
This is reflected on the fact that the perouter will be able to
establish two sessions with the host, one for ipv4 prefixes and one for
ipv6 prefixes.

Both will be translated to evpn type 5 routes.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>