Handle the emtpy local session scenario
When we don't have a local session, we don't need to:

- configure frr to establish a session
- create a veth pair to establish the session over the host with

Because of this, we split the setupVNI function in two, one where we
setup the VNI and one where we setup the veth to the host.

Additionally, we arbitrarily set the ASN of the router related to the
VNI's vrf to the underlay one. Since there is no session to be
established, it's arbitrary (and required by the frr configuration).
So, we choose the one we already have for the underlay. Exposing it to
the user is not intuitive as it's only needed to produce a valid frr
configuration.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>