clab: split the setup.sh script into modular pieces
Follow-up commits will introduce a new deployment testbed configuration,
along with a demo: it will provider 2 kind clusters, with an L2 overlay
network stretched between them.

Thus, to use that, we've had to split the setup.sh script in tinier
pieces, which will be executed independently. The argument to most of
the modular scripts is a list of kind cluster names.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>