cclm, multi-cluster: set up a dedicated migration network
In this commit, we set up a dedicated migration network in each cluster.
To minimize the configuration burden, we provision an additional L2VNI +
NAD in each cluster, and we use that network as the dedicated migration
network we provision in the `kubevirt` namespace.

We need to account for the VXLAN tunnel overhead, thus, we need to lower
the MTU to 1400 bytes.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>