e2etests: return error inside eventually block
The Expect inside Eventually prevents retries when bgpd isn't ready;
this change returns the error instead, which will enable the retry
logic.

Another option would be to receive `ginkgo` as an input parameter in
eventually, and run that `ginkgo` instance expectations. This would
allow the eventually block to be re-tried. I chose to follow the pattern
in place in this function, which is to return the errors.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>